fix memory allocator to align on 16 byte boundaries
Compilers like gcc assume malloc()'d memory is aligned to 16 byte
boundaries on 64 bit systems and may use SSE instructions that
require this to access structures.  Since the psmi_memtype_hdr
memory allocation metadata structure is prepended to malloc()'d
memory allocations it needs to be a multiple of 16 in size in
order preserve malloc alignment and avoid possible SEGVs when
SSE instructions are used (e.g. with -O3).

Resolve this by adding the original_allocation field from the
psm2 source tree to the header structure to make its size a
multiple of 16.  this allows us to easily add psm2's
psmi_memalign_internal() function in the future if we need it.
also, only set PSMI_STATSTYPE_MEMORY if PSM_DEBUG is enabled
to match psm2's behavior.